clean workflow models and admin site
